Starting to find my groove
Started creating native JS classes with inheritance and "static" methods.
Seemed to be race conditions with class definitions and imports, so I
switched back to the old style, abandoning the cleaner class syntax for
the uglier Object workflow and prototype syntax.

ResourceGroup seems to have evolved into my chokepoint for access control.
All/Most operations require an accessToken be passed into them.

Creating a new ResourceGroup requires a minor bit of bootstrapping to
bind the creator with an "owner" role without requiring an accessToken.

NOTE: Whenever an action creates a new ResourceGroup or otherwise modifies
a users permissions, they must have a new accessToken created and installed.

This would mean sending down a `Set-Cookie` header to the client browser
much like when we use the refreshToken to implicitly provide a new
accessToken to the user.